Move `BuildTracker` to `operations.build.build_tracker`
