+++ tests/unit/test_req.py
@@ -23,6 +23,7 @@ from pip._internal.exceptions import (
 from pip._internal.index.package_finder import PackageFinder
 from pip._internal.metadata.pkg_resources import Distribution
 from pip._internal.network.session import PipSession
+from pip._internal.operations.build.build_tracker import get_build_tracker
 from pip._internal.operations.prepare import RequirementPreparer
 from pip._internal.req import InstallRequirement, RequirementSet
 from pip._internal.req.constructors import (
@@ -39,7 +40,6 @@ from pip._internal.req.req_file import (
     get_line_parser,
     handle_requirement_line,
 )
-from pip._internal.req.req_tracker import get_build_tracker
 from pip._internal.resolution.legacy.resolver import Resolver
 from pip._internal.utils.urls import path_to_url
 from tests.lib import TestData, make_test_finder, requirements_file
